Refactored package names and added data binding to all the classes
