--- app/src/main/java/pulkit/com/torontowastewizard/Activity/Constants.java
@@ -1,7 +1,0 @@
-package pulkit.com.torontowastewizard.Activity;
-
-public class Constants {
-
-    public static String DOMAIN = "https://pulkitkumar.me/api";
-    public static String REGISTER = "/register.php";
-}
