+++ app/src/main/java/pulkit/com/torontowastewizard/activity/Common.kt
@@ -1,7 +1,9 @@
-package pulkit.com.torontowastewizard.Activity
+package pulkit.com.torontowastewizard.activity
 
 object Common {
 
     internal val name = arrayOf("bread bag tag", "milk bag tag", "elastic band", "rubber band", "twist tie", "rope", "twine", "string", "hemp", "ribbon", "bow", "burlap", "staple", "fastener", "wire", "florists wire", "plastic tag", "tape", "duct tape", "electrical tape", "masking tape", "scotch tape", "painters tape", "tape dispenser", "chain", "nylon", "thread", "metal lids", "glass bottle lids", "glass jar lids", "Mason jar lids", "laundry basket", "clothes basket", "hamper", "sewing machine", "vacuum cleaner", "carpet steamer", "water softener", "water softening system", "plastic fan", "ceiling fan", "molded paper", "paper egg cartons", "paper take-out coffee trays", "paper berry boxes", "fertilizer", "pesticide", "lawn chemicals", "Miracle Grow", "insect repellent", "insecticide", "herbicide", "round up", "round-up", "weed killer", "garden chemical", "2-4-D", "24D", "Diazinon", "weed spray", "plastic spray bottles", "cleaning spray bottles", "fireworks (unused)", "firecrackers (unused)", "pyrotechnics", "rockets", "sparklers", "illuminations", "road flares", "flares", "ammunition", "shrapnel", "bullets", "bullet shells", "projectiles", "scuba dive tanks", "cellulose nitrate films", "large ornament shells", "squeeze tube", "plastic hand cream tube", "plastic lotion tube", "plastic shampoo tube", "plastic sunscreen tube", "plastic sunblock tube", "plastic makeup tube", "plastic make-up tube", "empty metal tube", "toothpaste tube", "silicone plastic", "windshield washer fluid bottle", "antifreeze container", "car coolant container", "milk bags", "inner milk bags", "outer milk bags", "railway ties", "railroad tie", "railroad ties", "linoleum flooring", "rug", "mat", "rubber mat", "carpet", "carpet underlay", "undercarpeting", "underpadding", "hot plate", "motor (small)", "furnace motor", "metal range hood", "metal exhaust hood", "metal tire rim", "dehumidifier", "humidifier", "electronic air cleaner", "rototiller", "metal safe", "metal satellite dish", "antenna tower", "metal weight", "metal cooler", "steel bracket", "single-serve yogourt cups", "single-serve creamer containers", "single-serve milk containers", "single-serve jam packs", "single-serve peanut butter pack", "single-serve butter pack", "single-serve pudding cups", "single-serve fruit cups", "single-serve yogurt cups", "single-serve yoghurt cups", "books", "novel", "soft cover books", "paperback books", "hardcover books", "picture frame", "photo frame", "light fixture", "lamp", "solar lamp", "solar garden lamp", "solar garden light", "floor lamp", "juice drink pouch", "stand-up flexible pouch", "frozen fruit bag (stand-up pouch)", "frozen vegetable bag (stand-up pouch)", "retort pouch", "soap refill pouch", "Capri-Sun pouch", "beverage pouch", "Bibo", "Hi-C", "fruit juice pouch", "Kool-Aid Jammer", "tire", "tyre", "vehicle tire", "car tire", "small truck tire", "refrigerator", "fridge", "bar fridge", "large metal household appliance", "metal residential bar", "stove", "oven", "range", "cooker", "cooktop", "washing machine", "washer", "dryer", "drier", "clothes dryer", "dishwasher", "freezer", "deep-freeze", "freezer locker", "frozen food locker", "shower curtain", "shower liner", "shower caddy ", "metal chair", "metal table", "metal bed frame", "bed spring", "microwave oven", "metal fan", "exercise machine", "treadmill", "stationary bike", "rowing machine", "ironing board", "metal coat rack", "filing cabinet", "metal sink", "cast iron sink", "bathtub (metal)", "air conditioner", "metal vanity", "metal fireplace", "fireplace grate", "metal fireplace screen", "wood burning stove", "soap bottle (plastic)", "plastic shampoo bottle", "plastic lotion bottles", "plastic moisturizer bottles", "plastic sunscreen bottles", "plastic mouth wash bottles", "plastic hairspray bottles", "plastic cosmetic bottles", "plastic perfume bottles", "waxed cardboard box", "wax board", "waxed carton", "stretchy plastic appliance packaging", "tooth pick", "toothpick", "chewing gum", "nylons", "pantyhose", "panty hose", "water bed bladder", "plastic drum", "plastic barrel", "stuffed animal", "teddy bear", "doll", "chalk", "glow stick", "rubber dropper lid", "receipt (thermal/shiny)", "plastic tampon applicator", "plastic report cover", "clear plastic report cover", "empty gas can", "empty gas container ", "barbeque propane cylinder", "BBQ tank", "propane tank", "propane cylinder", "helium tank", "helium cylinder", "oxygen tank", "oxygen cylinder", "CO2 tank", "butane gas tank", "compressed gas tank", "pressurized gas cyclinder", "plastic take out food containers", "clam shell containers", "PETE #1 containers", "plastic produce containers", "plastic strawberry container", "plastic berry containers", "plastic bakery container", "plastic salad container", "plastic cake container", "plastic sprout container", "plastic tomato container", "clear plastic produce container", "plastic fruit container", "coffee cup lid", "plastic cup lid", "clear plastic clamshells", "household batteries", "AA battery", "AAA battery", "C battery", "D battery", "9V battery", "button cell battery", "alkaline-manganese battery", "lithium battery", "silver oxide battery", "zinc air battery", "zinc carbon battery", "lead acid battery", "CopperTop", "Duracell", "Energizer", "ni-cads", "rechargeable battery", "nickel cadmium battery", "linens", "textiles", "bedding", "bed sheets", "quilt", "duvet", "flags", "table cloth", "tablecloth", "towel", "microfibre cloth", "microfiber cloth", "cloth napkin", "landscape fabric", "cushion", "window covering", "blinds", "curtains", "drapes", "curtain rod", "plastic table cloth", "vinyl table cloth", "ear plugs", "foam ear plugs", "camera lens", "glue bottle (empty)", "glue stick (empty)", "ant trap", "garburator", "debit card", "credit card", "plastic gift card", "Metropass", "Glade air freshener", "solid air freshener", "slide projection carousel", "electric heating pad", "hot water bottle", "Plexiglas", "Acrylite", "Lucite", "Perspex", "Plexiglass", "black plastic bags", "black plastics", "hair", "human hair", "animal hair", "fur", "pet fur", "pet hair", "dog hair", "feather (real)", "razor", "disposable razor", "hair brush", "hair rollers", "hair comb", "dentures", "false teeth", "toothbrush", "tooth brush", "dental floss", "cast iron tub", "vehicle engine coolant", "ethylene glycol", "propylene glycol", "oil filter", "diesel fuel filter", "household furnace fuel filter", "coolant filter", "fuel", "diesel", "coal", "coal oil", "kerosene", "canned heat", "sterno", "fondue fuel", "alcohol", "petroleum", "lighter fluid", "camp fuel", "white gas", "camp fuel", "lamp oil", "gas", "gasoline", "paraffin oil", "motor oil", "motor oil jug", "motor oil bottle", "motor oil container", "pop cans", "soda cans", "aluminum cans", "beverage cans", "aluminum plates", "aluminum pie plates", "aluminum roasting pans", "rigid aluminum foil", "aluminum burner liner", "aluminum trays", "aluminum roasting trays", "biscuit tins", "cookie tins", "colostomy bag", "cast", "splint", "support", "catheter", "catheter bag", "catheter line", "dialysis waste", "dialysis tube", "dialysis bag", "gastric tube", "nasal tube", "IV bag", "IV tubing", "sponge (healthcare)", "bandage", "gauze", "Band-Aid", "intravenous tubing", "intravenous bag", "soiled dressing", "disposable mask", "disposable apron", "disposable gown", "disposable glove", "disposable pad", "disposable towel", "disposable sheet", "mesh produce bag", "plastic mesh bag", "plastic netting bag for fruits", "onions bag", "oranges bag", "clemenines bag", "mandarin bag", "burlap bag", "rice bag", "wooden crate", "clementines crate", "wooden basket", "wooden fruit basket", "construction papers", "Bristol boards", "posterboards", "paperboards", "cards stock", "cardstock", "cover stock", "pasteboards", "pumpkin", "gourd", "jack-o-lantern", "Halloween", "plastic flower pots", "plastic plant pots", "plastic plant trays", "plastic hanging baskets", "mercury thermostat", "mercury switch", "vacuum tube", "bear spray", "rat poison", "mouse poison", "gel pack", "ice pack", "gel ice pack", "hot pack", "cold pack", "hot cold pack", "cosmetics", "nail polish", "shoe polish", "shoe cream", "shoe paste", "roofing tar", "eye make-up", "mascara", "blush", "foundation", "lipstick", "Wite-out", "liquid paper", "fallen fruit from trees", "windfall", "walnuts", "pine cone", "pinecone", "acorn", "pine needles", "crab apples", "peat moss flower pot", "coconut flower pot", "coco-fiber", "coco-fibre", "planter", "wood finishing oil", "shellac", "varnish", "urethane", "block filler", "roof sealer", "driveway sealer", "foundation sealer", "paint aerosol", "spray paint", "aerosol can (full or partially full)", "sealant", "Asian longhorned beetle", "invasive species", "infected wood", "longhorn beetle", "frozen food box", "TV dinner", "microwave tray", "oven food tray", "plastic frozen food tray", "plastic food wrap", "food wrap", "Saran wrap", "Cling wrap", "shrink wrap", "stretch wrap", "meat wrap", "cheese wrap", "luncheon meat wrap", "vacuum sealed plastic wrap", "cheese slice wrapper", "aluminum foil", "alfoil", "alufoil", "tin foil", "waxed paper", "sandwich wrap", "parchment paper", "freezer paper", "plastic seal on margarine container", "fruit protection mesh", "sleeve netting", "foam packaging net", "pear wrap socks", "fruit foam sleeves", "fruit foam mesh", "mirror", "large mirror", "tall mirror", "oversized mirror", "floor mirror", "emerald ash borer beetle", "EAB", "invasive species", "infected wood", "fax machine", "facsimile machine", "fax copier", "photocopier", "copier", "Xerox copier", "typewriter", "car seat", "x-ray plate", "pool tarp", "pool cover", "pool liner", "vinyl pool liner", "hot tub cover", "pool blanket", "windsurf board", "solar panels (large)", "cooking oil (liquid)", "deep fryer oil", "vegetable oil", "pool filter", "food", "snack", "food scraps", "leftovers", "plate scrapings", "coffee grounds", "tea bag", "tea", "loose tea", "tea leaves", "Tetley tea", "Red Rose tea", "chicory", "egg", "eggshell", "nuts", "pecans", "filbert", "peanuts", "almonds", "pistachios", "cashews", "coconuts", "nut shells", "seeds", "candy", "chocolate", "candy cane", "fries", "french fries", "pencil shavings", "Paper straw (without wax/plastic lining)", "broken glass", "shard", "shattered glass", "broken window", "broken wine glass", "broken drinking glass", "cleaner", "detergent", "household cleaning product", "drain cleaner", "Draino", "oven cleaner", "Easy-Off", "chemical", "carpet cleaner", "car care product", "pool chemical", "chlorine puck", "chlorine", "adhesives (all types)", "glue", "Krazy glue", "Super glue", "epoxy", "acid", "base", "caustic chemicals", "muriatic acid", "blueprint ink", "polyethylene glycol solution", "Peglyte", "lubricant", "WD-40", "pepper spray", "compostable plastic", "biodegradable plastic", "plate", "compostable cutlery", "fork", "knives", "spoon", "cup", "degradable plastic", "bamboo cutlery", "bamboo plate", "sugarcane plate", "stand up pouch", "stand up flexible pouch", "tea light holders (aluminum only) ", "paint", "oil paint", "latex paint", "primer", "undercoat", "acrylic paint", "water-based paint", "alkyd paint", "enamel", "oil-based paint", "deck coating", "floor paint", "stain blocking paint", "textured paint", "concrete paint", "masonry paint", "drywall paint", "marine paint", "stain", "metal shovel", "metal snow shovel", "wine corks", "plastic wine corks", "wooden wine corks", "CD", "compact disc", "LP record", "eyeglasses", "eye glasses", "eyeware", "sun glasses", "sunglasses", "hearing aid", "walking aid", "television", "TV", "TV set", "VCR", "Video Cassette Recorder", "DVD player", "Digital Versatile Disk player", "Blu-ray player", "VHS player", "computer cables", "skewers", "wooden skewers", "BBQ skewers", "chopsticks", "wooden chopstick", "plastic straws", "plastic stir sticks", "lollipop stick", "beer caps", "beer bottle caps", "coffee disc", "Tassimo coffee pod", "Keurig coffee pods", "Nespresso coffee pods", "T-discs", "K cups", "k-cup", "coffee pods", "coffee pods", "coffee capsules", "nylon tea bag", "silk tea bag", "6-pack plastic rings", "12-pack plastic rings", "paper straw (with wax/plastic lining)", "beer bottles", "beer cans", "wine bottles", "wine cartons", "liquor bottles", "spirit bottles", "alcohol bottles", "gift wrapping paper", "tissue paper", "tissue wrapping paper", "plastic shopping bags", "polyethylene bags", "plastic grocery bags", "plastic retail bags", "tub", "bathtub (non-metal)", "fiberglass tub", "acrylic tub", "shower (dismantled)", "fiberglass shower stall", "shower insert", "metal shower enclosure", "shower walls", "jacuzzi", "vanity (non-metal)", "counter top", "countertop", "kitchen cabinet", "kitchen cupboards", "clothing", "clothes", "unwanted clothes", "footwear", "shoe", "runners", "sneakers", "boots", "sandals", "coat", "jacket", "dress", "hat", "scarf", "scarves", "shirt", "sweater", "pants", "jeans", "skirt", "purse", "hand bag", "umbrella", "life jacket", "underwear", "boxers", "book bag", "school bag", "backpack", "jewellery", "jewelry", "costume jewellery", "watch", "tie clips", "cuff links", "earrings", "necklace", "rings", "costume", "rags", "slipper", "slippers", "plastic outdoor garbage can", "blue box", "grey box", "gray box", "recycling box", "pizza box (greasy)", "paper plate (soiled)", "Chinet paper plate (soiled)", "paper bag", "brown Kraft bag", "cellulose lined paper bag", "paper food packaging (soiled)", "paper takeout container", "paper coffee filter", "muffin paper cup", "muffin wrapper", "cupcake paper cup", "cupcake wrapper", "paper flour bag", "paper popcorn bag", "fries box", "paper ice cream carton", "ice cream box", "BBQ charcoal bag", "electronic accessory", "coaxial cable", "headphones", "battery charger", "fibre optic cable", "fiber optic cable", "Mylar paper", "Mylar polyester film", "film", "polyester film", "static sheet", "matte", "overhead sheet", "plastic sheet", "plastic", "plastic paint tray", "plastic paint tray liner", "pool noodle", "foam pool noodle", "bathroom scale", "metal scale", "glass scale", "kitchen scale", "plastic (black)", "foam (black)", "cabinets", "shelf", "cubbie", "wardrobe", "wooden shelf", "book shelf", "bookshelf", "shelving", "wall unit", "entertainment unit", "hutch", "china cabinet", "medicine cabinet", "dog house", "doghouse", "wine barrel", "playpen", "wooden swing set", "meat (raw and cooked)", "poultry", "chicken", "beef", "pork", "burger", "steak", "ham", "bone", "gristle", "cartilage", "fish (raw and cooked)", "lobster", "seafood", "shellfish", "mollusc", "bivalve", "shrimp", "seafood shell", "mussels", "crab", "clam", "oyster", "periwinkles", "scallop", "snail", "bi-valves", "sea shells", "metal wheelbarrow", "metal drum", "metal barrel", "metal shed (dismantled)", "metal swing set", "metal hockey net frame", "metal mailbox", "yard waste", "wood chips", "wood shavings", "lawn chips", "mulch", "bark", "bark chips", "garden trimmings", "weeds", "small shrubs", "hay", "straw bales", "corn stalks", "outdoor garden plant", "plant bulb", "flower bulb", "lawn edging (earth removed)", "leaf", "leaves", "bamboo stick", "bamboo stake", "plastic cups", "polystyrene cups", "disposable plastic cups", "plastic cold drink cups", "Christmas tree (real)", "fruit", "produce", "fruit core", "fruit peels", "fruit stone", "fruit pit", "dried fruit", "rinds", "vegetable", "vegetable peels", "veggies", "lettuce", "salad", "salad dressing", "corn cob", "corn husk", "popcorn", "kernels", "small electric appliance", "small household appliance", "blender", "hair dryer", "curling iron", "blow dryer", "electric toothbrush", "paper shredder", "coffee maker", "kettle", "food processor", "juicer", "electric can opener", "electric razor", "indoor grill", "toaster", "toaster oven", "popcorn maker", "plastic bottles", "plastic water bottles", "plastic drink bottles", "plastic juice bottles", "plastic pop bottles", "plastic soda bottles", "PET bottles", "PETE bottles", "water cooler bottles", "plastic carboy", "plastic beverage containers", "plastic reusable water bottles", "Nalgene water bottles", "plastic paint pail", "plastic driveway sealant container", "plastic driveway tar container", "grass clippings", "thatch", "dead animal (small)", "dead mouse", "dead rat", "dead bird", "carcass", "remains", "laundry soap boxes", "dryer lint", "dryer sheets", "fabric softener sheets", "Bounce", "Downey", "Fleec", "anti-static sheets", "disposable mop sheets", "Swiffer", "disposable cleaning sheets", "duster", "household dusters", "vacuum cleaner bags", "vacuum contents", "sponges", "pot scrubber", "scouring pad", "steel wool", "cleaning pads", "SOS pads", "clothes hangers", "coat hangers", "wire hangers", "metal hangers", "plastic hangers", "coathangers", "wooden hangers", "bike", "bicycle", "tricycle", "plastic baby bottles", "metal baby bottles", "stainless steel baby bottles", "caulking", "sealant", "automotive grease", "car wax", "vehicle battery", "B8 car battery", "automobile battery", "motorcycle battery", "fire extinguisher", "mercury thermometer", "bird cage", "aquarium", "fish bowl", "fish tank", "aquarium sand", "Kitty litter (non-organic)", "crystal kitty litter", "silicone-based kitty litter", "silica gel kitty litter", "plastic water cooler jug", "plastic water cooler bottle", "15 litre water cooler bottle", "18 litre water cooler bottle", "5 gallon water cooler bottle", "8 litre water cooler bottle", "sour cream tubs", "margarine tubs", "plastic ice cream tubs", "cottage cheese tubs", "yogourt tubs", "plastic jars", "plastic food tubs", "snowblower", "snow blower", "lawnmower (gas", "electric or push)", "lawn mower", "oil filled heater", "radiator", "electric radiator", "oil-filled radiator", "oil tank (empty)", "garden brush", "twigs", "tree limbs", "branches", "hedge trimmings", "tree trimmings", "garden trimmings", "hedge clippings", "plastic jugs", "HDPE bottle", "plastic detergent bottles", "plastic bleach bottles", "plastic laundry detergent measuring cup", "glass bottles", "glass jars", "glass containers", "Mason jars", "glass carboy", "glass home brewing bottles", "glass wine making bottles", "glass cooking oil bottles", "baby food jar", "glass prescription bottles", "glass vitamin bottles", "glass perfume bottles", "paper lottery tickets", "packing paper", "kraft paper", "phone books", "telephone directory books", "file folders", "posters (unlaminated)", "paper greeting cards", "paper business card", "paper gift cards", "sticky notes", "Post-it notes", "paper postcards", "index cards", "receipt (matte paper)", "sales receipt (matte paper)", "cash receipt (matte paper)", "coffee cup sleeves", "diabetic test strips", "glucose meter", "glucometer", "inhaler (empty)", "Ventolin (empty)", "puffer (empty)", "pill blister pack", "home furnishings", "furniture", "bench", "stool", "wicker furniture", "bed frame", "mattress", "box spring", "sofa bed mattress", "sofa", "couch", "Chesterfield", "trunk", "chest", "recliner", "wooden chair", "plastic chair", "armchair", "La-Z-Boy recliner", "Lazy Boy", "smoke detectors", "smoke alarms", "alarms", "carbon monoxide detectors", "clocks", "watches", "clock radios", "alarm clock", "handheld calculators", "sex toys (any material)", "vibrators", "remote controls", "TV remote controls", "garage door opener remote", "flashlights", "vacuums", "electronic calculators", "adding machines", "tree stump", "large branch", "tree", "tree trunk", "string lights", "holiday lights", "Christmas lights", "festival lights", "light strings", "flower (artificial)", "flower (fake)", "plastic flower", "plastic plant", "silk flowers", "silk plants", "artificial plants", "florist foam", "Oasis floral foam", "party decorations", "balloons", "candle wax", "citronella candle", "election sign", "camera", "electronic thermometer", "personal electronic thermomter", "outdoor electronic thermometer", "electrical power bar", "e-book reader", "Kobo", "e-reader", "Kindle", "Treegator bag", "garden sprinkler", "clay flower pot", "clay planter", "garden hose", "hose", "rubber hose", "vinyl hose", "poisonous plant", "noxious plant", "invasive plant", "garlic mustard plant", "giant hogweed", "hogweed", "poison ivy", "dog strangling vine", "plastic shovel", "plastic snow shovel", "garden gloves", "work gloves", "rock salt", "road salt", "de-icer", "ice salt", "cardboard cans", "Pringles cans", "potato chips cardboard containers", "cardboard nut cans", "spiral cardboard cans", "cardboard coffee cans", "refrigerated dough cardboard containers", "Pillsbury dough cans", "powdered drink mix cardboard cans", "metal cans", "food cans", "tin", "steel cans", "soup cans", "coffee cans", "fish can", "tuna cans", "salmon cans", "sardines cans", "tomato sauce cans", "pasta sauce cans", "candy metal tins", "mint metal tins", "plastic pill bottles (empty)", "plastic vitamin bottles (empty)", "plastic medicine bottles (empty)", "plastic prescription bottles (empty)", "plastic medication vials (empty)", "sports equipment", "sporting goods", "soccer", "baseball", "baseball glove", "sports glove", "basketball", "golf ball", "football", "tennis ball", "bowling ball", "volleyball", "tennis racket", "badminton racket", "racquetball", "racket", "lacrosse stick", "racquet", "ice skates", "inline skates", "rollerblades", "roller skates", "figure skates", "hockey skates", "roller blades", "skateboard", "skates", "ink toner cartridges", "toner", "ink jet", "toner cartridges", "laser cartridges", "printer cartridges", "dirt", "sod", "fill", "grass", "sand", "soil", "rocks", "gravel", "patio stones", "paving stones", "cement", "grout", "cement powder", "stucco", "mortar", "clay pipe", "marble slab", "asphalt", "cinder block", "concrete pieces", "concrete block", "concrete sink", "concrete tub", "concrete laundry tub", "pool heater", "pool pump", "hot tubs", "baby gate", "toddler gate", "luggage", "suitcase", "baggage", "bundle buggy", "grocery cart", "water cooler refrigeration base", "water cooler", "refrigeration base", "popcorn cart", "Christmas tree (artificial)", "Christmas tree (fake)", "wheelchair", "ping pong table", "air hockey table", "plastic storage bin", "storage tote", "milk crate", "chalk board", "white board", "bulletin board", "workbench", "lawn furniture", "patio furniture", "patio umbrella", "patio table", "patio chairs", "plastic fence", "plastic shed (dismantled)", "rain barrel (empty)", "backyard composter (empty)", "Earth Machine composter", "ladder", "step ladder", "lawn roller", "lawn drum", "picnic table", "plastic swing set", "plastic slide", "plastic pool slide", "plastic frame hockey net", "baby swing", "splash pool", "children's pool", "plastic swimming pool", "trampoline", "plastic magazine overwrap", "plastic catalogue overwrap", "plastic flyer bag", "plastic newspaper wrapper", "toilet paper overwrap", "napkins overwrap", "paper towel overwrap", "water bottle overwrap (packaging)", "pop bottle overwrap (packaging)", "pop can overwrap", "soft drink overwrap (packaging)", "plastic overwrap", "overwrap", "over wrap", "diaper outer bag", "feminine hygiene outer bag", "pharmaceuticals", "medicines", "pills", "drugs", "prescriptions", "over the counter", "over-the-counter", "medication", "rubbing alcohol", "needles", "biohazardous", "hypodermic", "sharps", "lancets", "syringes", "vitamins", "supplements", "inhalers (full or partially full)", "puffers (full or partially full)", "Ventolin (full or partially full)", "medical injectors", "house plant (including soil)", "potted plant (including soil)", "dead plant (including soil)", "cactus (including soil)", "potting soil", "flower", "flower bouquet", "cut flowers", "dried flower", "electrical wire", "wiring", "cord", "extension cord", "fuse", "spark plug", "plastic light switch cover", "electronic adapters", "bricks", "asbestos", "PCBs", "polychlorinated biphenyls", "telephone pole", "tent trailer", "vending machine", "welding tank", "diving board", "pool table", "billiard table", "snooker table", "cue table", "piano", "organ", "oil tank (full or partially full)", "vehicle tank (full or empty)", "gas tank (full or empty)", "diesel tank (full or empty)", "vehicle propane tank (full or empty)", "fluorescent light bulb", "fluorescent light tube", "compact florescent light", "CFL", "compact fluorescent tube", "paper gift bags", "paper gift boxes", "boat", "motorized personal transporter", "sea-doo", "ski-doo", "jet ski", "snowmobile", "motorcycle", "scooter", "ebike", "electric bike", "segway", "golf cart", "plastic car parts", "pickup truck cabs", "truck cabs", "truck caps", "truck toppers", "plastic food tub lids", "plastic food jar lids", "Tupperware", "Rubbermaid food containers", "renovation materials", "lumber", "wood", "construction waste", "demolition waste", "C & D waste", "skid", "pressure treated wood", "ceiling tiles", "wood panelling", "plywood", "fencing", "baseboards", "quarter rounds", "plastic baseboard", "particle board", "particleboard", "chipboard", "Masonite", "aluminum siding", "vinyl siding", "shingles", "roof shingles", "saw dust", "sawdust", "gazebo", "awning", "garage door opener motor", "garage door rails", "mirror", "mirror (broken)", "coffee maker (broken glass)", "ceramic dish (broken)", "paring knife", "sharp knife", "utensil", "carving knife", "knives", "fork", "can opener", "nail", "screw", "nut", "bolt", "futon", "foam mattress topper", "stickers", "labels", "photographs", "photo", "pictures", "pic", "print", "x-ray", "photographic paper", "picture negatives", "photo negatives", "photo film", "blueprint paper", "blueprint", "carbon paper", "carbon tissue", "carbonic paper", "posters (laminated)", "laminated papers", "paper envelope with fibre cushioning", "scratch tickets", "scratch lottery tickets", "aerosol cans (empty)", "spray paint cans", "hairspray cans", "mousse cans", "compressed air cans", "shaving cream cans", "shaving gel cans", "shave gel cans", "shave cream cans", "whipping cream cans", "whipped cream cans", "insulation", "fibre-glass", "fibreglass", "pink fibre glass", "ductile foam sheet", "polystyrene rigid insulation", "Styrofoam insulation", "eavestrough", "downspout", "drywall (clean painted vinyl or paper-covered only)", "gyp rock", "plaster", "insulation urea formaldehyde", "coffee cup (takeout)", "coffee cup (paper)", "disposable coffee cup", "Dixie cup", "disposable hot beverage cup", "Tim Hortons cup", "Starbucks cup", "wax lined coffee cup", "wax coated coffee cup", "hot drink paper cup", "coffee cup (compostable)", "waxed paper cold drink cup", "sofa bed", "furnace filter", "air filter", "water filter", "Brita filter", "carbon water filter", "refrigerator water filter", "bitra filter", "plastic contact lenses cases", "plastic envelopes", "Canada Post plastic sealing envelopes", "plastic shipping envelopes", "plastic mailing envelopes", "plastic fruit baskets", "plastic berry baskets", "green plastic mesh fruit baskest", "dental floss boxes", "PETE 1 packaging", "plastic handles on cardboard baskets", "plastic cat litter boxes", "plastic kitty litter boxes", "plastic cat litter bins", "paper envelopes", "solvent", "degreaser", "chemical", "paint thinner", "paint stripper", "naphtha", "denatured alcohol", "mineral spirits", "turpentine", "lacquer thinner", "chian turpentine", "spirit of turpentine", "oil of turpentine", "wood turpentine", "gum turpentine", "vehicle engine antifreeze", "hockey stick", "skis", "downhill skis", "cross country skis", "snowboard", "golf club", "iron (golf)", "putter", "coat rack", "broom", "mop", "basketball stand (rim and pole)", "basketball hoop", "basketball pole", "wallpaper", "caulking tube", "caulk", "paintbrush", "paint brush", "rubber sheet", "drop sheet", "drop cloth", "vapour barrier", "tarp", "polyethylene plastic sheet", "plastic pipe", "plastic piping", "PVC pipe", "PEX pipe", "ABS pipe", "toilet seat", "drywall compound", "electric stove burner", "element", "ads", "flyers", "junk mail", "non-addressed mail", "newspapers", "magazines", "catalogues", "telephone", "phone", "cellular phone", "cell phone", "mobile phone", "smartphone", "iPhone", "Blackberry", "car phone", "PDA", "personal digital assistant", "pager", "answering machine", "audio electronics", "turntable", "radio", "receiver", "tuner", "stereo", "equalizer", "amplifier", "preamplifier", "preamp", "recording equipment", "stereo system", "compact disc player", "CD player", "Walkman", "speaker", "stereo speaker", "dairy product", "cheese", "curd", "brie", "pudding", "yogurt", "yoghurt", "yogourt", "sour cream", "dip", "cream", "ice cream", "cheese", "mozzarella", "ceramic sink", "plastic sink", "porcelain sink", "plastic laundry sink", "plastic laundry tub", "plastic egg cartons", "clear plastic food tubs", "clear plastic hummus containers", "clear plastic dip containers", "Ferrero Rocher containers", "plastic container for mushrooms", "polycarbonate food containers", "plastic cookie trays", "hard plastic meat trays", "plastic cracker trays", "plastic chocolate trays", "toiletries", "soap", "emery board", "deodorant", "antiperspirant", "deoderant", "baby wipe", "wet nap", "wet wipe", "towelette", "cleansing sheet", "wipes", "diaper wipe", "face wipe", "facial cleansing cloth", "make-up removing cloth", "make-up wipe", "make-up pad", "cosmetic pad", "", "ceramic vase", "cutting board", "wooden cutting board", "plastic cutting board", "metal cutlery", "barbeque", "BBQ", "BBQ grill", "barbecue", "milk cartons", "gable top containers", "soup cartons", "aseptic containers", "juice cartons", "juice boxes", "Tetra Pak containers", "soy milk cartons", "almond milk cartons", "soup cartons", "broth cartons", "cigarette", "cigarette butt", "ash", "tobacco", "fireplace ashes (cold)", "BBQ ashes (cold)", "charcoal (cold)", "lava rock (cold)", "briquette", "pyrotechnics (used)", "rockets (used)", "sparklers", "lighter (empty)", "Zippo lighter (empty)", "Bic lighter (empty)", "metal car parts", "muffler", "ride-on lawnmowers", "lawn tractor", "ride-on lawn mower", "cellophane", "bubbled plastic packaging", "bubble gum wrap", "bubble envelope", "bubblewrap", "metallic gift wrapping paper", "foam packaging peanuts", "packing peanuts", "packing foam chips", "foam packaging sheet", "silage wrap", "Instapak foam packaging", "calendars", "notebooks", "daytimers", "manual", "diary", "workbooks", "notepads", "spiral bound books", "facial tissue", "tissues", "Kleenex", "paper towel", "kraft paper towel", "paper napkin", "serviette", "diaper", "Pampers", "nappy", "Pull-ups", "Diaper Genie bag", "adult diaper", "puppy training pad", "incontinence pad", "Depends diapers", "feminine hygiene products", "tampon", "maxi pad", "pad", "sanitary napkin", "disposable absorbent pad", "Kotex", "Tampax", "sanitary liner", "insulin pens", "epinephrine auto injectors", "epipen", "insulin pen", "medical injector", "injector", "pre-filled pen", "pre filled pen", "bean bag chair", "beanbag chair", "dresser", "desk", "bureau", "bureaux", "table", "glass table", "crib", "stroller", "baby stroller", "baby carriage", "pram", "child seat", "baby car seat", "booster seat", "electronic greeting card", "musical greeting card", "electronic cigarette", "e cigarettes", "e-cigarettes", "hearing aid battery", "portable cell phone charger", "cell phone battery pack", "plastic reusable bag", "cloth reusable bag", "cotton reusable bag", "nylon reusable bag", "polyester reusable bag", "vinyl plastic bag", "reusable shopping bag", "reusable grocery bag", "lube", "WD40", "bleach", "Javex", "Clorox", "Ajax", "whitener", "photographic chemicals", "pool chemical", "algaecide", "pH buffer", "flea powder", "pet powder", "citronella oil", "hair dye", "hair colour", "perm solution", "nail polish remover", "lacquer", "baby oil", "hand sanitizer", "sunscreen", "sun tan lotion", "sunblock", "lotion", "moisturizer", "cream", "shampoo", "conditioner", "perfume", "cologne", "eau de toilette", "hydrochloric acid ", "music record", "audio tape", "LP", "video cassette", "computer disc", "floppy disc", "disc", "disk", "tape", "VHS", "beta", "DVD", "video game", "VHS tape", "sleep apnea breathing machine ", "frozen juice containers", "unwaxed cardboard", "old corrugated cardboard", "OCC", "corrugated fibreboard", "fiberboard boxes", "appliance boxes", "sauce", "au jus", "condiments", "gravy", "sugar", "spices", "herb", "honey", "syrup", "herbs", "grease", "fat (solid)", "lard", "butter", "margarine", "shortening", "mayonnaise", "Miracle Whip", "peanut butter", "spread", "jam", "jelly", "helmet", "trophy", "tent", "bicycle inner tube", "bike tire", "inner tube", "rubber inner tube", "playing cards", "deck of playing cards", "card deck", "plastic sandwich bags", "Ziploc bags", "baggie", "plastic storage bags", "reclosable plastic", "resealable plastic bags", "Ziploc sandwich bags", "Ziploc freezer bags", "plastic bread bags (non-foil)", "plastic produce bag", "produce bags", "plastic bulk food bags", "frozen vegetable bags (not stand-up pouch type)", "frozen fruit bags (not stand-up pouch type)", "expanded polystyrene foam", "metal plumbing fixture", "metal tap", "metal pipe", "showerhead", "shower head", "copper pipe", "faucet", "metal stake", "scrap metal", "wire fence", "chicken wire", "metal furnace", "hot water tank (empty)", "water heater", "baseboard heater", "pump", "water pump", "sewer pump", "sump pump", "laundry pump", "heat pump", "exchanger", "compressor", "tools", "power tool", "drill", "sander", "hammer", "screwdriver", "wrench", "pliers", "saw", "chisel", "toy (plastic or other material)", "holiday decoration", "wreath", "garland", "Christmas bulbs", "tinsel", "Christmas tree ornament", "decorations", "clothes pin", "clothes peg", "instrument", "drum (musical)", "matches", "manual typewriter", "sand filled weight", "plastic dumbbell", "picnic cooler", "plastic cooler", "condom", "rubber", "contraceptive", "Trojan condom", "Durex condom", "prophylactic", "condom", "rubber", "contraceptive", "Trojan condom", "Durex condom", "prophylactic", "latex rubber gloves", "disposable gloves", "rubber gloves", "paper clips", "nail clippings", "agricultural wrap", "boat wrap", "black garbage bag", "doors", "wooden door", "closet doors", "wooden garage doors", "folding wooden door", "metal doors", "aluminum doors", "metal garage doors", "screen door", "patio doors", "glass door", "shower doors", "window frames", "window glass", "window shutters", "wood flooring", "parquet flooring", "tiles", "ceramic tiles", "porcelain tiles", "slate tiles", "floor tiles", "paper", "writing paper", "lined paper", "fine paper", "printed paper", "computer paper", "fax paper", "glossy paper", "newsprint", "edge notched paper", "paper stock", "cereal", "grain", "oatmeal", "oats", "porridge", "gruel", "corn bread", "corn meal", "pasta", "bread", "flour", "baking supplies", "spelt", "kasha", "pasta", "noodles", "orzo", "couscous", "semolina", "rice", "Uncle Ben's rice", "baked good", "sweets", "dessert", "cake", "muffin", "cookie", "baking ingredient", "bread", "bun", "crust", "bread loaf", "pastry", "cupcake", "Timbit", "pie", "tart", "dough", "wheat", "pizza", "binders", "three ring binders", "bookbinders", "clipboards", "pens", "pencils", "writing instruments", "wooden pencils", "mechanical pencils", "markers", "marker pens", "marking pens", "crayons", "erasers", "cerlox binding", "comb binding", "surelox binding", "cork boards", "stapler (plastic)", "scissors (plastic)", "photo albums", "plastic squeezable food bottles", "plastic squeezable ketchup bottles", "plastic squeezable mayonnaise bottles", "plastic squeezable mustard bottles", "plastic squeezable salad dressing bottles", "plastic squeezable barbeque sauce bottles", "plastic squeezable soya sauce bottles", "plastic squeezable condiment bottles", "shredded paper", "shredded files", "shredded confidential documents", "pet food bag", "potato chip bag", "snack food bag", "paper chip bag", "cookie bag", "coffee bag", "candy wrapper", "chocolate wrapper", "granola bar wrapper", "butter wrapper", "yogurt tube", "plastic cereal liner bag", "plastic cracker box liner bag", "plastic cookie tray liner bag", "pasta bag", "blister packaging (gum and lozenges)", "butcher paper (for wrapping meat)", "pad from meat tray", "meat tray liner", "chocolate box pad", "metal paint cans (empty)", "metal paint can lids", "roofing tar metal pails (empty)", "roofing tar cans (empty)", "dead animal (large)", "carcass", "corpse", "remains", "dead raccoon", "thermoform packaging", "clear plastic packaging", "clear molded plastic packaging", "blister packaging", "plastic coin wrapper", "paperboard", "cardboard barrel", "cardboard boxes", "pizza boxes (empty - no food residue or plastics)", "cereal boxes", "cracker boxes", "pasta boxes", "boxboard", "tissue boxes", "Kleenex boxes", "shoe boxes", "cigarette packages", "pet food", "nibbles", "dog food", "cat food", "rawhide chews", "dog treat", "dog bone", "bird seed", "bird feed", "kitty litter (organic)", "animal bedding", "cat litter", "pet litter", "clay litter", "clay kitty litter", "pet waste", "excrement", "feces", "animal droppings", "doggie doo", "dog pooh", "dog poo", "dog poop", "pet diaper", "puppy diaper", "dog diaper", "DVD cases", "CD cases", "CD jewel cases", "Cafe", "light bulb (incandescent or LED)", "incandescent light bulb", "halogen light bulb", "pot light bulb", "flood light bulb", "light-emitting diode", "light emitting diode", "LED", "solar panels (small)", "plastic pails", "plastic drums", "plastic buckets", "metal bottles", "aluminum bottles", "stainless steel bottles", "reusable metal bottles", "metal water bottles", "metal mugs (without any plastic)", "reusable stainless steel mugs (without any plastic)", "metal cooking oil containers", "polystyrene foam food packaging", "coffee cup (styrofoam)", "foam cup", "takeout foam coffee cup", "foam egg carton", "polystyrene foam cup", "foam plate", "foam takeout container", "foam take-out container", "Styrofoam takeout container", "foam tray (non-black)", "foam meat tray (non-black)", "foam clamshell container", "Styrofoam clamshell container", "extruded foam packaging", "electronic foam packaging", "styrofoam plate", "plastic bags (soft and stretchy)", "polyethylene plastic bags", "dry cleaning bags", "polyethylene dry cleaning bags", "mulch bags", "sand bags", "triple mix bags", "fertilizer bags", "soil bags", "peat moss bags", "manure bags", "compost bags", "road salt bags", "clear recycling bags", "transparent recycling bags", "mattress bags", "plastic newspaper bags", "housewares", "dishes", "plastic dish rack", "drinking glass", "cup", "mug", "metal cookware", "pot", "pan", "kitchen utensil", "flatware", "flatware tray", "frying pan", "baking tray", "Teflon pan", "plate", "tableware", "Pyrex", "stoneware", "china", "visionware", "earthenware", "pottery", "ceramic dishes", "glass dish", "vase", "glass vase", "porcelain vase", "crystal vase", "art and craft materials", "art supplies", "button", "sequin", "bead", "feather (fake)", "zipper", "magnet", "key", "plasticine", "playdo", "modelling clay", "play doh", "playdough", "crafts", "wooden popsicle stick (clean)", "yarn", "threads", "cotton fibre", "fiber", "flaxen thread", "spun wool", "board game", "puzzle", "camp tarp", "paper rolls", "toilet paper rolls", "paper towel rolls", "toilet paper tubes", "paper tubes", "cardboard tubes", "toilet", "toilet bowl", "toilet tank", "toilet set", "one piece toilet", "bidet", "electronic waste", "computer", "laptop", "PC", "desktop computer", "computer tower", "portable computer", "monitor", "CRT monitor", "LCD monitor", "computer notebook", "computer notepad", "computer terminal", "modem", "computer peripheral", "computer speaker", "disk drive", "computer accessories", "keyboard", "mouse", "speaker", "headset", "printer", "Q-tip", "cotton-tipped swab", "ear swab", "ear cleaner", "cotton swab", "cotton buds", "qtip", "cotton ball", "cotton swab", "talcum powder", "baby powder", "detergent box", "laundry detergent box", "dish detergent box", "metal floor register", "metal floor vent", "licence plates", "", "salad spinner", "office tray ", "Goose down duvet", "pillows ", "disposable plastic plates", "plastic cutlery", "plastic forks", "plastic knife", "plastic spoons", "fireworks (used)")
+    val DOMAIN = "https://pulkitkumar.me/api"
+    val REGISTER = "/register.php"
 
 }
