+++ app/src/main/java/pulkit/com/torontowastewizard/activity/FavActivity.kt
@@ -1,43 +1,42 @@
-package pulkit.com.torontowastewizard.Activity
+package pulkit.com.torontowastewizard.activity
 
+import android.databinding.DataBindingUtil
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import android.view.WindowManager
-import androidx.databinding.DataBindingUtil
 import java.util.ArrayList
-import pulkit.com.torontowastewizard.Adapter.WasteAdapter
-import pulkit.com.torontowastewizard.Model.Waste
+import pulkit.com.torontowastewizard.adapter.FavWasteAdapter
+import pulkit.com.torontowastewizard.model.Waste
 import pulkit.com.torontowastewizard.R
-import pulkit.com.torontowastewizard.databinding.ActivitySearchBinding
+import pulkit.com.torontowastewizard.databinding.ActivityFavBinding
 
-class SearchActivity : AppCompatActivity() {
+class FavActivity : AppCompatActivity() {
 
-    companion object {
-        private val TAG: String = SearchActivity::class.java.simpleName
-    }
+    private val TAG: String = FavActivity::class.java.simpleName
 
-    private lateinit var binding : ActivitySearchBinding
+    private lateinit var binding : ActivityFavBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_search)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_fav)
 
         window.setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN)
 
         val wasteArrayList = intent.getSerializableExtra("wasteArrayList") as ArrayList<Waste>
-        val adapter = WasteAdapter(wasteArrayList)
         val layoutManager = LinearLayoutManager(this)
         val itemDecor = DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
         itemDecor.setDrawable(this.resources.getDrawable(R.drawable.line_divider))
+        val adapter = FavWasteAdapter(wasteArrayList)
 
-        binding.wasteRecyclerView.apply {
+        binding.wasteRecyclerViewFav.apply {
             this.layoutManager = layoutManager
             addItemDecoration(itemDecor)
             this.adapter = adapter
         }
+
     }
 }
