+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activity.SearchActivity"
+        tools:context=".activity.SearchActivity"
         android:background="@drawable/bg">
 
         <androidx.recyclerview.widget.RecyclerView
